Revert "Merge branch 'develop' into sh/dcj-755-staging-testrunnersa"
This reverts commit 05e7bb07ac48da7ce8ce149d3b7f0262168f7092, reversing
changes made to 48e5cd65a8be25850ac8609bbcdbaba2e1113bc0.
