+++ README.md
@@ -29,13 +29,13 @@ See [started guide](docs/jade-getting-started.md) for information on running con
 ### Run TDR locally
 
 To run TDR locally:
-`./scripts/run start_local`
+`./scripts/run local`
 
 To run TDR in docker:
-`./scripts/run start_docker`
+`./scripts/run docker`
 
 To run TDR locally and wait for debugger to attach on port 5005:
-`./scripts/run start_local --debug-jvm`
+`./scripts/run local --debug-jvm`
 
 To have the code hot reload, enable automatic builds in intellij, go to:
 `Preferences -> Build, Execution, Deployment -> Compiler`
