+++ build.gradle
@@ -20,7 +20,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath('io.swagger.codegen.v3:swagger-codegen:3.0.64')
+        classpath('io.swagger.codegen.v3:swagger-codegen:3.0.63')
         // Required for gradle liquibase plugin
         classpath ('org.liquibase:liquibase-core:4.29.2')
     }
@@ -33,7 +33,7 @@ plugins {
     id 'org.gradle.test-retry' version '1.6.0'
     id 'antlr'
     id 'org.hidetake.swagger.generator' version '2.19.2'
-    id 'org.springframework.boot' version '3.4.0'
+    id 'org.springframework.boot' version '3.3.5'
     id 'idea'
     id 'java'
     id 'io.spring.dependency-management' version '1.1.6'
@@ -44,13 +44,13 @@ plugins {
     id 'com.dorongold.task-tree' version '4.0.0'
     // enables release info in sentry events
     id 'com.gorylenko.gradle-git-properties' version '2.4.2'
-    id 'org.sonarqube' version '6.0.1.5171'
+    id 'org.sonarqube' version '5.1.0.4882'
     id 'com.srcclr.gradle' version '3.1.12'
 }
 
 allprojects {
     group 'bio.terra'
-    version '2.198.0-SNAPSHOT'
+    version '2.182.0-SNAPSHOT'
 
     ext {
         resourceDir = "${rootDir}/src/main/resources/api"
@@ -63,8 +63,8 @@ allprojects {
             mavenBom SpringBootPlugin.BOM_COORDINATES
         }
         dependencies {
-            dependency 'io.swagger.core.v3:swagger-annotations:2.2.26'
-            dependency 'io.swagger.codegen.v3:swagger-codegen-cli:3.0.64'
+            dependency 'io.swagger.core.v3:swagger-annotations:2.2.25'
+            dependency 'io.swagger.codegen.v3:swagger-codegen-cli:3.0.63'
         }
     }
 }
@@ -165,14 +165,14 @@ dependencies {
     implementation 'com.google.apis:google-api-services-iam:v1-rev20230209-2.0.0'
     implementation 'com.google.auth:google-auth-library-oauth2-http:1.19.0'
 
-    implementation platform('com.google.cloud:libraries-bom:26.51.0')
+    implementation platform('com.google.cloud:libraries-bom:26.50.0')
     implementation 'com.google.cloud:google-cloud-billing'
     implementation 'com.google.cloud:google-cloud-resourcemanager'
     implementation 'com.google.cloud:google-cloud-bigquery'
     implementation 'com.google.cloud:google-cloud-storage'
     implementation 'com.google.cloud:google-cloud-firestore'
     implementation 'com.google.cloud:google-cloud-pubsub'
-    implementation 'com.google.cloud:spring-cloud-gcp-starter-logging:5.9.0'
+    implementation 'com.google.cloud:spring-cloud-gcp-starter-logging:5.8.0'
     implementation 'com.google.http-client:google-http-client'
 
     implementation 'org.apache.commons:commons-compress:1.27.1' // For srcclr, jib plugin conflict
@@ -202,14 +202,14 @@ dependencies {
     implementation 'com.microsoft.sqlserver:mssql-jdbc:12.3.0.jre17-preview'
 
     // For distributed locking of Spring @Scheduled tasks across multiple instances
-    implementation 'net.javacrumbs.shedlock:shedlock-provider-jdbc-template:6.0.2'
-    implementation 'net.javacrumbs.shedlock:shedlock-spring:6.0.2'
+    implementation 'net.javacrumbs.shedlock:shedlock-provider-jdbc-template:5.16.0'
+    implementation 'net.javacrumbs.shedlock:shedlock-spring:5.16.0'
 
     implementation 'bio.terra:terra-common-lib:1.1.18-SNAPSHOT'
     implementation 'org.broadinstitute.dsde.workbench:sam-client_2.13:v0.0.241'
     implementation 'bio.terra:terra-policy-client:1.0.17-SNAPSHOT'
     implementation 'bio.terra:terra-resource-buffer-client:0.198.42-SNAPSHOT'
-    implementation 'bio.terra:externalcreds-client-resttemplate:1.54.0-SNAPSHOT'
+    implementation 'bio.terra:externalcreds-client-resttemplate:1.52.0-SNAPSHOT'
 
     implementation 'org.glassfish.jersey.inject:jersey-hk2'
 
@@ -222,26 +222,26 @@ dependencies {
     implementation 'com.fasterxml.jackson.core:jackson-databind'
 
     // Azure related dependencies
-    implementation 'com.azure:azure-identity:1.14.2'
-    implementation 'com.azure.resourcemanager:azure-resourcemanager:2.45.0'
+    implementation 'com.azure:azure-identity:1.14.0'
+    implementation 'com.azure.resourcemanager:azure-resourcemanager:2.44.0'
     implementation 'com.azure.resourcemanager:azure-resourcemanager-loganalytics:1.0.0'
     implementation 'com.azure.resourcemanager:azure-resourcemanager-securityinsights:1.0.0-beta.5'
-    implementation 'com.azure:azure-storage-common:12.28.0'
-    implementation 'com.azure:azure-storage-file-datalake:12.22.0'
-    implementation 'com.azure:azure-data-tables:12.5.0'
+    implementation 'com.azure:azure-storage-common:12.27.1'
+    implementation 'com.azure:azure-storage-file-datalake:12.21.0'
+    implementation 'com.azure:azure-data-tables:12.4.6'
 
-    implementation platform('io.sentry:sentry-bom:7.18.1') //import bom
+    implementation platform('io.sentry:sentry-bom:7.16.0') //import bom
     implementation('io.sentry:sentry-spring-boot-starter-jakarta')
     implementation('io.sentry:sentry-logback')
 
     // OpenTelemetry @WithSpan annotations:
-    implementation 'io.opentelemetry.instrumentation:opentelemetry-instrumentation-annotations:2.10.0'
+    implementation 'io.opentelemetry.instrumentation:opentelemetry-instrumentation-annotations:2.8.0'
 
-    testImplementation 'org.apache.parquet:parquet-common:1.15.0'
-    testImplementation 'org.apache.parquet:parquet-hadoop:1.15.0'
-    testImplementation 'org.apache.parquet:parquet-hadoop-bundle:1.15.0'
-    testImplementation 'org.apache.parquet:parquet-encoding:1.15.0'
-    testImplementation 'org.apache.parquet:parquet-column:1.15.0'
+    testImplementation 'org.apache.parquet:parquet-common:1.14.3'
+    testImplementation 'org.apache.parquet:parquet-hadoop:1.14.3'
+    testImplementation 'org.apache.parquet:parquet-hadoop-bundle:1.14.3'
+    testImplementation 'org.apache.parquet:parquet-encoding:1.14.3'
+    testImplementation 'org.apache.parquet:parquet-column:1.14.3'
     testImplementation ('org.apache.hadoop:hadoop-common:3.4.1')  {
         exclude group: 'com.sun.jersey', module: 'jersey-core'
         exclude group: 'com.sun.jersey', module: 'jersey-servlet'
@@ -257,8 +257,8 @@ dependencies {
         exclude group: 'org.slf4j', module: 'slf4j-reload4j'
     }
 
-    testImplementation 'au.com.dius.pact.provider:junit5:4.6.16'
-    testImplementation 'au.com.dius.pact.provider:junit5spring:4.6.16'
+    testImplementation 'au.com.dius.pact.provider:junit5:4.6.15'
+    testImplementation 'au.com.dius.pact.provider:junit5spring:4.6.15'
 
     antlr 'org.antlr:antlr4:4.13.2'
 
@@ -268,8 +268,8 @@ dependencies {
 
     testImplementation 'org.junit.vintage:junit-vintage-engine'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'io.zonky.test:embedded-database-spring-test:2.6.0'
-    testImplementation 'io.zonky.test:embedded-postgres:2.1.0'
+    testImplementation 'io.zonky.test:embedded-database-spring-test:2.5.1'
+    testImplementation 'io.zonky.test:embedded-postgres:2.0.7'
     implementation enforcedPlatform('io.zonky.test.postgres:embedded-postgres-binaries-bom:12.8.0')
 
     generatedCompile 'org.springframework.boot:spring-boot-starter-web'
@@ -277,7 +277,7 @@ dependencies {
     generatedCompile 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.webjars:webjars-locator-core'
-    runtimeOnly 'org.webjars.npm:swagger-ui-dist:5.18.2'
+    runtimeOnly 'org.webjars.npm:swagger-ui-dist:4.3.0'
     generatedCompile 'io.swagger.core.v3:swagger-annotations'
 
     annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
