+++ src/main/java/bio/terra/common/PdaoLoadStatistics.java
@@ -1,6 +1,5 @@
 package bio.terra.common;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.cloud.bigquery.JobStatistics;
 import java.time.Instant;
 
@@ -10,11 +9,7 @@ public class PdaoLoadStatistics {
   private final Instant startTime;
   private final Instant endTime;
 
-  public PdaoLoadStatistics(
-      @JsonProperty("badRecords") long badRecords,
-      @JsonProperty("rowCount") long rowCount,
-      @JsonProperty("startTime") Instant startTime,
-      @JsonProperty("endTime") Instant endTime) {
+  public PdaoLoadStatistics(long badRecords, long rowCount, Instant startTime, Instant endTime) {
     this.badRecords = badRecords;
     this.rowCount = rowCount;
     this.startTime = startTime;
