+++ src/main/java/bio/terra/service/dataset/DatasetSchemaUpdateValidator.java
@@ -12,12 +12,15 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
 
 @Component
 public class DatasetSchemaUpdateValidator implements Validator {
+
+  @Autowired private DatasetRequestValidator datasetRequestValidator;
 
   @Override
   public boolean supports(Class<?> clazz) {
@@ -26,9 +29,10 @@ public class DatasetSchemaUpdateValidator implements Validator {
 
   private void validateDatasetSchemaUpdate(DatasetSchemaUpdateModel updateModel, Errors errors) {
     if (DatasetSchemaUpdateUtils.hasTableAdditions(updateModel)) {
-      SchemaValidationContext context = SchemaValidationContext.forUpdate();
+      DatasetRequestValidator.SchemaValidationContext context =
+          new DatasetRequestValidator.SchemaValidationContext();
       for (TableModel tableModel : updateModel.getChanges().getAddTables()) {
-        context.validateTable(tableModel, errors);
+        datasetRequestValidator.validateTable(tableModel, errors, context);
       }
       List<String> newTableNames = DatasetSchemaUpdateUtils.getNewTableNames(updateModel);
       Object[] duplicateTables =
