+++ src/main/java/bio/terra/service/dataset/flight/create/CreateDatasetRegisterIngestServiceAccountStep.java
@@ -1,20 +1,15 @@
 package bio.terra.service.dataset.flight.create;
 
-import bio.terra.common.exception.ErrorReportException;
 import bio.terra.service.auth.iam.IamService;
 import bio.terra.service.dataset.flight.DatasetWorkingMapKeys;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.FlightMap;
 import bio.terra.stairway.Step;
 import bio.terra.stairway.StepResult;
-import bio.terra.stairway.StepStatus;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /** The step is only meant to be invoked for GCP backed datasets. */
 public class CreateDatasetRegisterIngestServiceAccountStep implements Step {
-  private static final Logger logger =
-      LoggerFactory.getLogger(CreateDatasetRegisterIngestServiceAccountStep.class);
+
   private final IamService iamService;
 
   public CreateDatasetRegisterIngestServiceAccountStep(IamService iamService) {
@@ -27,18 +22,7 @@ public class CreateDatasetRegisterIngestServiceAccountStep implements Step {
     String datasetServiceAccount =
         workingMap.get(DatasetWorkingMapKeys.SERVICE_ACCOUNT_EMAIL, String.class);
 
-    try {
-      iamService.registerUser(datasetServiceAccount);
-    } catch (ErrorReportException e) {
-      // This is the super class type of IamExceptions (i.e. IAmNotFoundException) and ApiExceptions
-      // Catch transient errors from Sam, where the service account created in the previous step
-      // may not be ready to use yet. Do not catch InterruptedExceptions.
-      logger.warn(
-          String.format(
-              "Service account, %s, may not be ready to use yet. Retrying.", datasetServiceAccount),
-          e);
-      return new StepResult(StepStatus.STEP_RESULT_FAILURE_RETRY, e);
-    }
+    iamService.registerUser(datasetServiceAccount);
 
     return StepResult.getStepResultSuccess();
   }
