+++ src/main/java/bio/terra/service/dataset/flight/create/DatasetCreateFlight.java
@@ -102,9 +102,7 @@ public class DatasetCreateFlight extends Flight {
       // Create the service account to use to ingest data and register it in Terra
       if (datasetRequest.isDedicatedIngestServiceAccount()) {
         addStep(new CreateDatasetCreateIngestServiceAccountStep(resourceService, datasetRequest));
-        addStep(
-            new CreateDatasetRegisterIngestServiceAccountStep(iamService),
-            getDefaultExponentialBackoffRetryRule());
+        addStep(new CreateDatasetRegisterIngestServiceAccountStep(iamService));
       }
     }
