+++ src/main/java/bio/terra/service/filedata/azure/tables/TableDirectoryDao.java
@@ -440,7 +440,7 @@ public class TableDirectoryDao {
                 // Store the batch of entries. This will override existing entries,
                 // but that is not the typical case and it is lower cost just overwrite
                 // rather than retrieve to avoid the write.
-                storeDirectoryEntries(snapshotTableServiceClient, snapshotId, snapshotEntries);
+                batchStoreDirectoryEntry(snapshotTableServiceClient, snapshotId, snapshotEntries);
 
                 return null;
               }));
@@ -497,16 +497,5 @@ public class TableDirectoryDao {
                         () ->
                             createEntityForPath(tableClient, snapshotId, tableName, snapshotEntry)))
             .toList());
-  }
-
-  // Store the directory entries in the snapshot table sequentially, without a thread pool
-  void storeDirectoryEntries(
-      TableServiceClient snapshotTableServiceClient,
-      UUID snapshotId,
-      List<FireStoreDirectoryEntry> snapshotEntries) {
-    String tableName = StorageTableName.SNAPSHOT.toTableName(snapshotId);
-    TableClient tableClient = snapshotTableServiceClient.getTableClient(tableName);
-    snapshotEntries.forEach(
-        snapshotEntry -> createEntityForPath(tableClient, snapshotId, tableName, snapshotEntry));
   }
 }
