+++ src/main/resources/application.properties
@@ -145,6 +145,7 @@ azure.retryTimeoutSeconds=3600
 rbs.enabled=false
 rbs.poolId=datarepo_v1
 rbs.instanceUrl=https://buffer.tools.integ.envs.broadinstitute.org
+rbs.clientCredentialFilePath=/tmp/buffer-client-sa-account.json
 ecm.basePath=https://externalcreds.dsde-dev.broadinstitute.org
 ecm.rasIssuer=https://stsstg.nih.gov
 usermetrics.appId=datarepo
