+++ src/test/java/bio/terra/service/dataset/flight/update/DatasetSchemaUpdateValidationTest.java
@@ -61,7 +61,7 @@ import org.springframework.test.context.ContextConfiguration;
     })
 @WebMvcTest
 @Tag(Unit.TAG)
-class DatasetSchemaUpdateValidateModelStepTest {
+class DatasetSchemaUpdateValidationTest {
 
   @MockBean private JobService jobService;
   @MockBean private DatasetService datasetService;
