+++ src/test/resources/application-connectedtest.properties
@@ -29,3 +29,4 @@ db.migrate.dropAllOnStart=true
 rbs.enabled=true
 rbs.poolId=datarepo_v1
 rbs.instanceUrl=https://buffer.tools.integ.envs.broadinstitute.org
+rbs.clientCredentialFilePath=/tmp/buffer-client-sa-account.json
