--- src/test/resources/update-schema-bad-data-type.json
@@ -1,25 +1,0 @@
-{
-  "description":"Test invalid data type",
-  "changes":{
-    "addTables":[
-      {
-        "name":"new_table",
-        "columns":[
-          {
-            "name":"bad_column",
-            "datatype":"badDataType",
-            "array_of":false,
-            "required":false
-          }
-        ],
-        "primaryKey":null,
-        "partitionMode":"none",
-        "datePartitionOptions":null,
-        "intPartitionOptions":null,
-        "rowCount":null
-      }
-    ],
-    "addColumns":null,
-    "addRelationships":null
-  }
-}
