+++ tools/rbs-bearer-token.sh
